fixed curatino

+++ general_utils/byol_utils.py
@@ -7,4 +7,5 @@ def _create_model_training_folder(writer, files_to_same):
         os.makedirs(model_checkpoints_folder)
         for file in files_to_same:
             os.makedirs(os.path.join(model_checkpoints_folder, os.path.split(file)[0]))
-            copyfile(file, os.path.join(model_checkpoints_folder, os.path.split(file)[0]))
+            if not(os.path.exists(os.path.join(model_checkpoints_folder, os.path.split(file)[0]))):
+                copyfile(file, os.path.join(model_checkpoints_folder, os.path.split(file)[0]))
